requirements.txt
torch>=2.0.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.21.0
fastapi>=0.100.0
uvicorn>=0.23.0
redis>=4.5.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
databases[mysql]>=0.8.0
pymysql>=1.0.0
python-multipart>=0.0.6
python-dotenv~=1.0.1
joblib~=1.4.2
pydantic~=2.10.6
requests~=2.32.4
selenium==4.15.0
html-testRunner==1.2.1
webdriver-manager==4.0.1